Remove old v4 Channel 🙌 - Adapter & error imporvements in adapters
