+++ primitives/src/util/tests/prep_db.rs
@@ -1,24 +1,21 @@
 use crate::{
     campaign::{self, Active, Validators},
-    channel::{Pricing, PricingBounds},
-    channel_v5::Nonce,
+    channel::Nonce,
     targeting::Rules,
-    AdUnit, Address, BigNum, Campaign, Channel, ChannelId, ChannelSpec, EventSubmission,
-    SpecValidators, UnifiedNum, ValidatorDesc, ValidatorId, IPFS,
+    AdUnit, Address, Campaign, Channel, ChannelId, EventSubmission, UnifiedNum, ValidatorDesc,
+    ValidatorId, IPFS,
 };
 use chrono::{TimeZone, Utc};
-use fake::faker::{Faker, Number};
 use hex::FromHex;
 use lazy_static::lazy_static;
 use std::{collections::HashMap, convert::TryFrom};
 
 lazy_static! {
-    // dummy auth
-    // session_tokens
+    /// Dummy [`crate::ValidatorId`]s
     pub static ref IDS: HashMap<String, ValidatorId> = {
         let mut ids = HashMap::new();
 
-        ids.insert("leader".into(),  ValidatorId::try_from("0xce07CbB7e054514D590a0262C93070D838bFBA2e").expect("failed to parse id"));
+        ids.insert("leader".into(), ValidatorId::try_from("0xce07CbB7e054514D590a0262C93070D838bFBA2e").expect("failed to parse id"));
         ids.insert("follower".into(), ValidatorId::try_from("0xc91763d7f14ac5c5ddfbcd012e0d2a61ab9bded3").expect("failed to parse id"));
         ids.insert("user".into(), ValidatorId::try_from("0x20754168c00a6e58116ccfd0a5f7d1bb66c5de9d").expect("failed to parse id"));
         ids.insert("publisher".into(), ValidatorId::try_from("0xb7d3f81e857692d13e9d63b232a90f4a1793189e").expect("failed to parse id"));
@@ -29,6 +26,7 @@ lazy_static! {
         ids
     };
 
+    /// Dummy [`crate::Address`]es
     pub static ref ADDRESSES: HashMap<String, Address> = {
         let mut addresses = HashMap::new();
 
@@ -43,7 +41,7 @@ lazy_static! {
         addresses
     };
 
-    // These are the Goerli testnet  Addresses of these stablecoins
+    // These are the Goerli testnet [`Addresses`] of the following stablecoins:
     pub static ref TOKENS: HashMap<String, Address> = {
         let mut tokens = HashMap::new();
 
@@ -53,8 +51,9 @@ lazy_static! {
         tokens
     };
 
-    // dummy auth tokens
+    // Dummy adapter auth tokens
     // authorization tokens
+    ///
     pub static ref AUTH: HashMap<String, String> = {
         let mut auth = HashMap::new();
 
@@ -110,35 +109,7 @@ lazy_static! {
         }
     };
 
-    pub static ref DUMMY_CHANNEL: crate::channel::Channel = {
-        let nonce = BigNum::from(<Faker as Number>::between(100_000_000, 999_999_999));
-
-        crate::channel::Channel {
-            id: ChannelId::from_hex("061d5e2a67d0a9a10f1c732bca12a676d83f79663a396f7d87b3e30b9b411088").expect("prep_db: failed to deserialize channel id"),
-            creator: ValidatorId::try_from("033ed90e0fec3f3ea1c9b005c724d704501e0196").expect("Should be valid ValidatorId"),
-            deposit_asset: "0x89d24A6b4CcB1B6fAA2625fE562bDD9a23260359".to_string(),
-            deposit_amount: 1_000.into(),
-            targeting_rules: Rules::new(),
-            // UNIX timestamp for 2100-01-01
-            valid_until: Utc.timestamp(4_102_444_800, 0),
-            spec: ChannelSpec {
-                title: None,
-                validators: SpecValidators::new(DUMMY_VALIDATOR_LEADER.clone(), DUMMY_VALIDATOR_FOLLOWER.clone()),
-                max_per_impression: 10.into(),
-                min_per_impression: 1.into(),
-                targeting_rules: Rules::new(),
-                event_submission: Some(EventSubmission { allow: vec![] }),
-                // July 29, 2019 7:00:00 AM
-                created: Utc.timestamp(1_564_383_600, 0),
-                active_from: None,
-                nonce: Some(nonce),
-                withdraw_period_start: Utc.timestamp_millis(4_073_414_400_000),
-                ad_units: vec![],
-                pricing_bounds: Some(PricingBounds {impression: None, click: Some(Pricing { max: 0.into(), min: 0.into()})}),
-            },
-            exhausted: Default::default(),
-        }
-    };
+    pub static ref DUMMY_CHANNEL_ID: ChannelId = ChannelId::from_hex("061d5e2a67d0a9a10f1c732bca12a676d83f79663a396f7d87b3e30b9b411088").expect("prep_db: failed to deserialize channel id");
 
     pub static ref DUMMY_AD_UNITS: [AdUnit; 4] = [
         AdUnit {
