+++ sentry/src/db/channel.rs
@@ -95,7 +95,7 @@ pub async fn insert_validator_messages(
 
 mod list_channels {
     use primitives::{
-        channel_v5::Channel,
+        channel::Channel,
         sentry::{channel_list::ChannelListResponse, Pagination},
         ValidatorId,
     };
