+++ sentry/src/db/event_aggregate.rs
@@ -2,7 +2,7 @@ use chrono::{DateTime, Utc};
 use futures::pin_mut;
 use primitives::{
     balances::UncheckedState,
-    channel_v5::Channel as ChannelV5,
+    channel::Channel as ChannelV5,
     sentry::{EventAggregate, MessageResponse},
     validator::{ApproveState, Heartbeat, NewState},
     Address, BigNum, ChannelId, ValidatorId,
